move users componetns into components folder
