+++ client/config/Routes.jsx
@@ -2,15 +2,12 @@ import React from 'react';
 import { Router, Route, IndexRoute, hashHistory } from 'react-router';
 import App from '../../client/components/App.jsx';
 import SearchTest from '../../client/components/SearchTest.jsx';
-import Main from '../components/Main.jsx';
-import UserForm from '../users/UserForm.jsx';
-import SignUp from '../components/SignUp.jsx';
-import Login from '../components/Login.jsx';
+import UserForm from '../components/users/UserForm.jsx';
 
 const Routes = () => (
   <Router history={hashHistory} >
     <Route path="/" component={App} >
-      <IndexRoute component={SearchTest}/>
+      <IndexRoute component={SearchTest} />
       <Route path="signup" component={UserForm} />
       <Route path="login" component={UserForm} />
     </Route>
